+++ src/Application/Interface.ts
@@ -5,9 +5,6 @@
  * @library Application/Interface
  * @includes IConfig Application/_Interface/IConfig
  * @includes IEnv Application/_Interface/IEnv
- * @includes ISerializableState Application/_Interface/ISerializableState
- * @includes IStore Application/_Interface/IStore
- * @includes IRequest Application/_Interface/IRequest
  * @includes IHead Application/_Interface/IHead
  * @includes IHttpRequest Application/_Interface/IHttpRequest
  * @includes IHttpResponse Application/_Interface/IHttpResponse
@@ -20,10 +17,10 @@
 export * from 'Application/_Interface/IConfig';
 export * from 'Application/_Interface/IEnv';
 // export * from 'Application/_Interface/ILocation';
-export * from 'Application/_Interface/ISerializableState';
-export * from 'Application/_Interface/IStore';
+// export * from 'Application/_Interface/ISerializableState';
+// export * from 'Application/_Interface/IStore';
 // export * from 'Application/_Interface/IStateReceiver';
-export * from 'Application/_Interface/IRequest';
+// export * from 'Application/_Interface/IRequest';
 export * from 'Application/_Interface/IHead';
 export * from 'Application/_Interface/IBody';
 export * from 'Application/_Interface/IHttpRequest';
