+++ src/Application/_Env/NodeJS/Env.ts
@@ -1,10 +1,10 @@
 /// <amd-module name="Application/_Env/NodeJS/Env" />
 import { Config } from 'Application/Config';
 import {
-    IEnv, IHttpRequest, IHttpResponse,
-    IRequest, IRequestInternal
+    IEnv, IHttpRequest, IHttpResponse
 } from 'Application/Interface';
 import Request from 'Application/Request';
+import type { IRequest, IRequestInternal } from 'Application/Request';
 import type { IConsole } from 'Application/_Env/IConsole';
 import type { ICookie, ILocation } from 'Application/_Env/Interfaces';
 import Console from 'Application/_Env/NodeJS/Console';
